Add recursion if the db fails to return a success
I've seen couch/cradle sometimes fail to throw to an error, but return
a null response and fail to save a record. This might be due to
iriscouch's timeout problems. At any rate, this introduces a loop of
saving to the DB until we get back a success.

It's totally possible that this could cause an infinite loop. Yup.