Convert to using `init`
*This is a breaking change.*

`options` no longer accepts a callback method. It, now accepts it as
part of the the normal `app.use(plugin, options, callback)` syntax.

This also moves db init to the app init. This means that the server will
start before a db connection is made. Perhaps slightly slower, but this
makes it easier to cooperate with other plugins in the flatiron universe